fix(lint): clippy-wasm warnings in Multikicker implementation (mtg-914)
Two categories of clippy lint failures caught by make validate lint.clippy-wasm:

1. `useless_conversion` (7 instances): `u8::from(card.times_kicked)` and
   `u8::from(kick_cost.cmc())` where the inner type is already `u8` — `ManaCost::cmc()`
   returns `u8` and `times_kicked` is `u8`. Removed all `u8::from()` wrappers.

2. `wildcard_match` (1 instance): the `_ =>` catch-all arm in `apply_etb_counters`'s
   `CountExpression` match. Replaced with an exhaustive explicit list of all remaining
   variants (ValidPermanents, CardsDrawnThisTurn, CardsInHand, XPaid, TargetedCardPower,
   SpellsCastThisTurn, ValidGraveyard, Compare, Kicked, Bargain) so future variant
   additions require explicit handling here.

No behavior change — pure lint cleanup. `cargo clippy -p mtg-engine --features network
--all-targets -- -D warnings` now clean.

+++ mtg-engine/src/game/game_loop/priority.rs
@@ -1184,7 +1184,7 @@ impl<'a> GameLoop<'a> {
                                             }
                                         });
                                     if let Some(kick_cost) = multikicker_cost {
-                                        let kick_cmc = u8::from(kick_cost.cmc());
+                                        let kick_cmc = kick_cost.cmc();
                                         if kick_cmc > 0 {
                                             self.mana_engine.update_mut(self.game, current_priority);
                                             let untapped_sources =
